requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-bootstrap-datepicker-plus==5.0.0
django-bootstrap4==22.2
django-mdeditor==0.1.20
django-widgets-improved==1.5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
Markdown==3.4.1
oauthlib==3.2.2
openpyxl==3.0.10
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.11.0
